requirements.txt
hyperliquid-python-sdk